removed some superfluous things

+++ SKIRTcore/ProcessAssigner.cpp
@@ -67,13 +67,13 @@ bool ProcessAssigner::validIndex(size_t absoluteIndex) const
 
 ////////////////////////////////////////////////////////////////////
 
-std::vector<int> ProcessAssigner::indicesForRank(size_t rank) const
+std::vector<int> ProcessAssigner::indicesForRank(int rank) const
 {
     std::vector<int> result;
     result.reserve(_nvalues);
 
     // add all the absolute indices that correspond to the given rank
-    for (int absoluteIndex=0; absoluteIndex<total(); absoluteIndex++)
+    for (size_t absoluteIndex=0; absoluteIndex<total(); absoluteIndex++)
         if (rank == rankForIndex(absoluteIndex))
             result.push_back(absoluteIndex);
